iirfilter: move the c->gain multiplication from the input side of the filter to the output side.
This is a better match to the typical implementation and makes the state
values have the correct scale so that the output coefficients do not need
to be scaled. This was only done previously in order to normalize the output
coefficients to integers for high pass and low pass filters.